backend/stt-service/requirements.txt
fastapi
openai-whisper
uvicorn
python-multipart
numpy
torchaudio
torch